Made a some kind of AI. Firstly tried to implement it with setTimeout and setInterval, but due to it's asynchrony things got messy. Function calls weren't synchronized and nothing worked well. So I changed the approach and used Promises.
Now throw animation is made with Promise in which we have setInterval, where every 100 ms is called rollDiceAnimation.
Then we generate a random dice and depending on which we either make a delay of 0.5s to show user that he rolled 1, after that reset all the round points, or sum new ones. In both cases it calls resolve with random rolled dice.

So when it's AI's turn we call aiTurn and feed it 1. That means it's the first throw.
Then it calls aiThrow function which returns a promise. Inside we check if amount of throws isn't greater that 4 and round points aren't greater 15. If it's so, we callRoll dice function and set a timeout to make computer moves a bit slower.
After rollDice function is executed, we check which value it has returned.
Greater that 1? Calling resolve with increased amount of throws.
1? Calling reject with code 1.
Also if the initial check isn't passed reject is called, but now with code 0.

After that, we go back to aiTurn and depending on result decide what to do next:
If the promise return resolve than we make a delay and after that throw dice again.
Otherwise we check the code of reject.
If it's 1 -- do nothing. Everything has already been done in rollDice (reseting points and turn passing).
If 0 -- call holdPoints function.

Besides now pressing buttons roll dice and hold doesn't directly call rollDice and holdPoints.
rollDiceIfPossible and holdPointsIfPossible check if user can perform those actions and do everything that's needed to prevent unexpected behaviour.

That's all. Seems to work as swiss watch, but who knows.